Began implementing classes.
This might be easier than I thought.  I was hung up on forcing the classes to use __init__ statements, when I realized that
for the time being, they really don't need them.  I've converted storage_buffer and edge_transition into classes so far,
and the inheritance of the log file and operation time from the facility class seems to be working great thus far.
And since the code is already calling functions from modules, it's a fluid transition of turning modules into classes
and having the code run the same method calls.

I'm thinking of revamping the naming scheme: of just using all lower case letters for the component module and class names.
Because normally, classes have capitalized letters to differentiate them, but I think it will be rather obvious what
are classes once everything is such.

My next step is to simply go through and throw everything into classes similar to how I've done these two so far and
see if the code runs.  Using the inhereted operation time increment seems to work, but it needs to be universally implemented
to make the code pass the unit test without constantly reassigning the operation time being passed to modules from the
operation time being stored in the classes.

This is looking even greater than expected so far.  It's already solved a problem I was having ealier, and I think
the inheritance will significantly streamline the code once I implement all the io functions.  But, again, next
step is to make the components into objects and make sure it at least passes the unit test.