Successfully converted all of the used modules into classes
I've changed up the naming scheme of the modules that contain the classes, but I kept all of the old functions just in case.
Either way, I changed the operation time to be contained in facility_vars that will get passed to every method.  Eventually I want it to be the case
where everything that needs to get passed from component to component will be contained in either facility_vars, or a "batch" class.

That being said, now that I have a proof of concept for the new architecture integrated with the previous code, I can go about eliminating the old
function variables one by one and begin storing them in the class that most makes sense.  I've almost done this with operation time, all that's left is to
delete the old operation time variable and rely entirely on the operation time stored in facility_vars.

Again, once that is done, I can go after the other variables one by one.  The biggest question at this point is how to integrate all of the io functions and the myriad
of directories for the input and output, but I think they, too, can simply be integrated into the existing classes (at this point, I feel like most everything will go into
facility_vars.  Although that wouldn't be bad, because the code would run faster by passing classes than to pass the dozen or so variables back and forth).