Added batch module
Added an extra class that I want to be initialized in the
storage buffer.  It will then get passed from vertex to vertex
and get acted on by each one if needs be.  It will then
carry the relevant information to product storage.

I've noticed a monstrous bug with the measured storage
inventory.  It's not even close to what it should be.  I don't
know if that's in the original code, or if I inadvertently created
it.  Either way, I'm going to bring in the original code and
save the output files that it creates.

After I track down that bug, I'm heavily considering just
re-writing mainflow from scratch.  After dealing with the code
thus far, I feel I've got a really good handle on the vision for
such.  At this point, I really feel that simply re-transcribing
a new mainflow would be faster and easier than changing the
existing one to incorporate the object-oriented shift.