Extracted original output
I was able to run the original code and extract the output
files to my desktop.  I will now get those back in my master
branch as the Unit Test Docs to be compared to.

Indeed the bug with the measured storage inventory existed in
the original code, but everything else seems to be working
the way it's supposed to.  Now that I have the original log
and other output files, I going to start rewriting mainflow.