Rewrote mainflow
So, this is a big one.  I created a new directory entirely
to house my idea, but I went in and completely rewrote
the fuel fabrication mainflow.  It's not entirely complete,
as I still have to loop over what I have so far.  But as
of thus far, it does everything in the log file that the
first mainflow does.

Everything except for the end of campaign inspection is sleeker
and more streamlined than the original code in my opinion.
The end of campaign inspection can be cleaned up by doing
some preliminary calculations before writing the output
to the log file, though.

The only thing that's left to get the mainflow up to where it was
before I started is to include the output files, but
that should be able to get integrated very easily into the
architecture I have so far.

After that's done, implementing false alarm and failure routines
into the system should be able to done much more quickly than
before, since now everything is wrapped up in classes instead
of passing dozens of separate variables back and forth.