Began designing melter failure routine
I changed it so that the melter accumulates a heel object that
will be able to be passed from object to object once I design such.

For now, I implemented the weibull distribution calculation inside
the melter (although I think it'd be better suited inside the
facility component module, I will move it later) and with such
made a method that the melter could use to create an equipment
failure.  Although I have yet to code what happens when an
equipment failure actually occurs.

Also, I bundled up everything that actually belongs to the
fuel fabrication process inside a new class appropriately dubbed
fuel_fabricator.  Now mainflow works more like what the end goal is:
it brings in the facility components which themselves have components
and passes the SNM from component to component.

All that I really should have left at this point is implementing
the failure routine and then implement alarm systems.

As a final comment: note that in the weibull distribution, eta
is equal to the failure rate, not the inverse of it.  This
corresponds with dimensional analysis as well as what my own
research has shown the properties of the cumulative
distribution function to be.  And that is that eta
is the point at which the distribution reaches 60%.