Completed fail routine and modified kmps
******Short Version**************
I completd the melter failure routine.

Expected weight values were moved to being a state variable of
fuel_fabricator and product_storage.

Moved equipment fail check to facility_component.

Replaced unit test log.

**************Long Version*****************
I shook things up quite a bit while finishing the fail routine.
I decided that the expected weight values shouldn't be something
inside of batch, because the actual batches aren't going to
have any computer components that facility personnel could program
to decide/estimate how much material is going to be lost during
each operation.  So I decided to make that a variable of the
facility components instead.  I didn't put it in the
overarching facility class, because I figured that each variable
should be pushed as far down the totem pole as possible for
modularity's sake.

I did my best to simulate the failure routine according to the
papers I read, although it might have to be changed later.

Obviously once that was implemented, the old unit test became
useless.  So I did as much debugging as I could to make sure
that the system was doing simply what it sounds like it should
be doing.  As of thus far, one gets negative MUF values once
the melter fails and the heel is cleaned out, but I'm leaving
this as is for now simply to represent the recovery of once lost
materials.  Maybe this should get changed, but I'm going to keep
it as is for the moment.

That being said, I replaced the unit test with the most recent
log entry.  At this point I'm going on a bit of faith that
the code is actually working and everything is functioning
the way it's supposed to.

Assuming the failure routine is actually coded correctly, the
next step would be to implement an alarm system into the KMP's
and end of campaign inspection.

How an alarm should be set off is straight forward enough.
Exactly what happens when an alarm is set is something
I'm yet to be sure of.