Overhauled edge_transition and expected_weight
********Short Version************

Made edge transition shuttle expected batch weights from vertex
to vertex as well as "move" the batch itself.

Created an expected_weight_class that gets inherited by all
components that aids in keeping track of the expected batch
weight.

***************Long Version***************

In light of trying to come up with system alarms and facility
inspections, I figured it'd be best to find a more robust way
of dealing with the expected weight.  Instead of my original idea
of having facility and fuel_fabrication micromanage the expected
weight, I implemented a system where each component gets the
expected weight of the incoming batch from the component before
it.  And if the state variables change along the way, than the
component in charge of that state variable modifies the expected
weight as well.

I've been doing my best to modularize the code in a
hierchical decomposition.  So now the storage units keep track
of their own true, expected, and measured weights instead of
something else doing such for them.  This is all thanks to the
heavily modified edge transition, which can pass th relevant
information to such.

I believe my code now still works as before without any bugs,
but I'm not entirely sure.  Next I will do a more thorough scan
of how the expected weight is being handled, as some sporadic code
had to be tossed into the melter cleaning function, the product
storage, and the recycle storage in order to make things work right.
I will clean up that code and make sure everything is working
as it should before moving onto making a better end of campaign
and facility inspection.

My vision for that is that the facility will call on its
immediate children, then if those children have any children,
they will inspect their children, and so on and so forth.  That
way an ordered hierarchy is created with respect to reporting
given values.