Finished tweaking expected weight and end of campaign inspection
*******************Short************************

Added extra methods to expected_weight module

Stripped edge transition to just pushing expected batch weight

Added method to kmp to report weight to a given storage buffer

Modified end of campaign inspection to appear similar to
original

*****************Long**************************
The short version pretty much sums it up.

I've been struggling with automation vs. customizability.
The edge transition, for example, took care of a lot more, but
I'm afraid that putting to many things into one object would
spoil the readability and understandability of the code.

The flip side of that is that several different methods have
to be called everytime a state variable changes or a measurement
is made (but only when next to a storage unit).

I'm doubting the structure of my code more and more.  However,
at this point, everything is now set up to create reliable
weight checks at the kmp's and at the end of a campaign.  So I'll
do that and implement an alarm system.

But first, I'm going to go back and change command and control
to be as flexible as it's supposed to be before my meeting
with Dr. Borrelli.