Modified command and control
****************Short*********************

Edited the command and control main function to be able to edit
the "global_vars" file in a specified location.

Fixed a minor bug that came up in recycle_storage

******************Long*********************

Command and control should now be fully operational as
originally desired.  If done correctly, when the code is
freshly forked onto a new computer, one simply has
to run the command and control main function, input the desired
directories as prompted, and then everything should get taken
care of from there.

This was accomplished by making command and control rewrite
the "global_vars" file used by the specified module (fuel.fabriction,
oo.fuel.fabrication, etc.).

After seeing that the program runs 226 days essentially just as
quickly as 10 days, I found it reasonable and better to just keep
the original operation time.  This revealed a bug in
recycle_storage that was quickly stamped out.

Now that that is accomplished.  The next task is to finally
finish the code by implementing alarms and inspections.  Then,
in theory, I should be done. Except for any sort of data outputting
that might be warranted.