Created Facility inspection
**************SHORT*************
Created facility inspection method that updates measured and
expected weights in storage unit inventories

*************LONG********************
I'm about to restructure the code so that the facility will
contain all of its own components.  This is something I always
considered doing, but now it has become necessity.

This is the only way that the components of the facility can
effectively "talk up" to it.  I ran into this problem
while trying to figure out how to make the kmp alarm
routine.  The end of campaign routine works fine since
one is already in an environment where everything is passed in.
But passing the batch from a random kmp to the recycle storage
isn't possible with things set up the way they are now.

I rigorously checked the log file as it works thus far, and it
seems to be good.  So I copied it over for this next part.