Restructured facility command and main flow
***********SHORT**************
Moved all the facility components into the facility
itself

**************LONG***************
That didn't take nearly as long as I expected, which is good.
But I got all the components under the facility command
module.  Now the hierchical decomposition of the code is more
complete, creating a logic tree of sprawling branches that
represent the entire facility.  The branches can, should, and do
speak laterally to each via edge transitions, but now they can
speak directly to the head as well without having to pass
information directly.

Now I should be able to go in and write a proper kmp alarm
routine.  One that will break the normal batch processing
routine, move the batch to recycle storage, then conduct
a facility inspection as already laid out in the end of
campaign alarm routine.