Created kmp alarm
***************SHORT*********************
Rearranged end of campaign inspection, created facility
material accounting, and made kmp alarm routine.

***************LONG*********************
What used to be in the end of campaign inspection is now
its own routine with added optional arguments that run
a slightly different version of materials accounting if
a kmp is passed in.

I'm saving this now, because currently, the kmp threshold is set
to be too large to ever get set off.  But I'm about to toy with
such to see how the program runs when an alarm is set off.