Completed and ran kmp alarm routine
*****************SHORT***************
Made two kmps detect discrepancies in the alloted time

**************LONG********************
I changed the threshold for kmps 1 and 2 so that it detects at
least one weight discrepancy.  I ran the program, and it seemed
to change significantly from the unit test (at least there are
a LOT fewer end of campaign alarms).  I'm going to have to look
into this next.