Made minor tweaks finishing kmp alarm routine
************SHORT*********************
Tweaked output of kmp alarm.  Rigorously checked old with new log
file to make sure the changes caused by the new routine didn't
create any discrepancies

****************LONG***************************
The changes were minor and few and are summarized well enough in
the short description.

The next step is simply to add time delays to the inspection
routines.  Once that is done, the code should be ready to
use in calculating the final, publishable results.