Completed alarm routines and updated comments
*************SHORT*****************
Everything works as it should and is well
documented.  The code is essentially good to go

*************LONG******************
As said above, everything functions as it should as
far as I can tell.  All that's left it to write the appropriate
output files depending on what kind of results we want
to make graphs out of.