Redid alarm routine
******************SHORT************************
Renamed certain labels in the log file

Refit alarm routine to not readjust expected and measured
weights.

Alarm routine now moves cleaned heel back into the storage
buffer.

******************LONG*************************
I had to readjust the threshold for the end of campaign mass
check, because 2 was way too small without readjusting expectations
at some point or another.  Also, I scrapped the idea of kmp alarms
for the time being for that very reason.

However, I guess that false alarms occurring so frequently is the
point of this simulation, but it is going so far as to falsely
detect diversion attempts as well, because the disparity between
measured and actual weights accumulates overtime.  I need
to find a pleasing way to fix this.

In the mean time, I'm going to leave things as they are and
figure out how to make a campaign muf and output the data files
the way they were in the procedural version of the code.