Added campaign MUF
******************SHORT************************
Included campaign MUF calculator in fuel fabrication.

Added flags in facility command to facilitate this.

******************LONG*************************
It was about as tricky as expected with the way things are set
up, but I made it work.  Calculating the measured MUF was the worse
(as expected) since the hierarchy doesn't necessarily make
it smooth for communication to jump between units aside from
normal edge transitions.

But using a couple new flags in facility command (melter_did_fail
and did_conduct_inspection), I was able to make it work.

Next step is to create data output files.