Added data output
******************SHORT************************

Wrote and added data_output_module to facility component class
as well as melter and facility.

******************LONG*************************

In order to create data output routines, I went the route of
my handling of expected weight in the sense I added a new class
with routines specific to each case.  This might be clunky and
bloated, but it's the best solution I could come up with at the time.

To the best of my knowledge, I now have data output files for
everything that the procedural program did.

The next steps are to label the columns within the files (they
can still be read-in using np.loadtxt using the "skiprows" option).
Then I need to look over command and control, because it not
being able to handle directories that are already written is a bit
annoying.  But more importantly, it is not correctly creating all
the output directories that should have been created.

Lastly, perhaps I will include the check that the melter will
clean itself after accumulating 2 kg of heel to avoid critical
mass with the batch.  And after that, we should be able to create
data of how the probability of false alarms affects operation time.

WAIT!  I really need to create a false alarm and failure counter.
Those are really important.