Included remeasurement routine in inspection
******************SHORT************************

Added routines in facility inspection where the product storage
and storage buffer get updated measured inventories.

******************LONG*************************

To circumnavigate the problem of accumulated error in the
storage units continually setting off alarms, I included
a small routine where the storage buffer and product storage
undergo a "personal inspection" by facility personnel.  This
remeasures the measured inventory of each to more accurately
portray the true inventory.  Something akin to this should
be done in a real facility inspection anyways.

The next steps to take from here are all listed on the previous
commit.  But on top of that, reworking my slopped in routines
to be more bundled up would be a good idea.  Also, I should bring
in the inspection measurement uncertainty in from input files.
Then command and control will have to modified again.