Cleaned up new inspection and began editing command and control
******************SHORT************************

Bundled up inspection routines to utilize hierarchy

Moved mainflow to the root directory and began changing command
and control.

******************LONG*************************

The inspection routines now actually go through the proper
management units.  This doesn't do much now, but in the
future, calling these from a list will make things easier
and smoother.

I made another input file for the measurement uncertainty
of the inspection of the storage units.

I moved mainflow into the root directory.  This will act as
the "executable" that will dig into the source directory to
do everything that should be done.  I've started tweaking
little things in command and control, but the big thing
to do is simply create a new simulation directory, run mainflow,
and see from there what's missing in lib and/or what got
disconnected in all the editing.

Global vars might be redundant at this point, I should find
a way to eliminate it.  The only problem is deciding which
simulation directory to use.