Edited lib and tweaked command and control
*************SHORT****************
Edited "lib" directory so that new inputs used in changes are
included whenever a new user sets up a new simulation.

Changed command and control to account for existing simulation
directories.

*************LONG*****************
The short version basically says it all.  A couple things that
I still want to tweak before making plots and wrapping this thing
up for presentation are:

1. Label the input files and change the mainflow to skip the
label lines.  A short description for each would've gone a long
way for me.

2. Although it'd be a bit much to overall command and control
to keep it from changing directories so much, at the very least
make it so that it winds back up in the root directory.

3. Tweak command and control to change the simulation directory
even when the root directory isn't changed.