Added basic plot routines
************SHORT*****************
Made a general and specific plot methods in facility_command

*************LONG*****************
I did as mentioned.  Note that I made a "data_output" routine,
although I feel it is antiquated and comes from the very first
part of the code design, so it might need looking over and
over-hauled.  Either way, I've started the first steps towards
getting plots to work.

Also, don't forget the things that need to be done as listed
in the May 16 commit.