Added failure and false alarm counters to bolster plots
************SHORT*****************
Made two plots to display number of failure and false
alarm counters

*************LONG*****************
The code is ready to produce some results.  I want to change
the failure rate, measurement uncertainty, and discrepancy to
see how it changes the false alarm rate.

Also, don't forget the things mentioned in the last commit.